Implemented Basic Health and Mana Bars
Basic Health and Mana bars done.

TODO:
Need barUI design
Fiddle with values (do we want it to be a % or actually show the real value on the bar too e.g. 55/100)
Do we want to include the exp bar on the top panel as well?
Might need to include a few more methods to be called at other places to update the displays?